requirements.txt
# Web Framework
Django
django-environ
django-crispy-forms
crispy-bootstrap5

# Database
psycopg2-binary

# PDF Processing
pdfminer.six
python-magic
pdf2image

# OCR
pytesseract
Pillow

# Vector Database
pinecone-client
# Graph Database
neo4j

# Text Processing & ML
nltk
scikit-learn
sentence-transformers
spacy

# File Storage
django-storages
boto3

# Utilities
python-dateutil
tqdm

# Development & Testing
pytest
pytest-django
black
flake8
coverage


google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
celery
redis
celery[redis]